Move one dependency code to an internal module
The `classic-ancii-tree` package has not been updated for a long time and has a very old version of the `chalk` package as a dependency that pulls vulnerable dependencies.

At the same time, the latest version of `chalk` is already used in the project. And the `classic-ancii-tree` code is a small function. And it's easier to move it to the project, where it will use the already available `chalk`.